+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Utilitário `format_voter_id` [#221] (https://github.com/brazilian-utils/brutils-python/issues/221)
+- Utilitário `get_address_from_cep` [#358](https://github.com/brazilian-utils/brutils-python/pull/358)
+- Utilitário `get_cep_information_from_address` [#358](https://github.com/brazilian-utils/brutils-python/pull/358)
+- Utilitário `format_voter_id` [#221](https://github.com/brazilian-utils/brutils-python/issues/221)
 - Utilitário `generate_voter_id` [#220](https://github.com/brazilian-utils/brutils-python/pull/220)
 
 ## [2.1.1] - 2024-01-06